move: DashboardRepository 패키지 이동(#23)
