+++ src/main/java/com/dnd/namuiwiki/domain/dashboard/DashboardRepository.java
@@ -1,4 +1,4 @@
-package com.dnd.namuiwiki.domain.statistic;
+package com.dnd.namuiwiki.domain.dashboard;
 
 import com.dnd.namuiwiki.domain.statistic.model.entity.Dashboard;
 import com.dnd.namuiwiki.domain.survey.type.Period;
@@ -8,6 +8,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.Optional;
 
-public interface StatisticsRepository extends MongoRepository<Dashboard, String> {
+public interface DashboardRepository extends MongoRepository<Dashboard, String> {
     Optional<Dashboard> findByUserAndPeriodAndRelation(User user, Period period, Relation relation);
 }
